feat: Add --skip-legacy-booting-support-flag
On systems with UEFI support the grub dependency and install process are
not necessary. Add a flag to skip the grub dependency check and install
process.

This patch also moves the runtime dependency checking to be *after* the
commandline parameter processing since the grub runtime dependency is only
required when we're not skipping legacy boot support.

Fixes #109

Signed-off-by: Mitchel Humpherys <mitch.special@gmail.com>